Misc organization things
Time to start cleaning up
