+++ TISFAT/UI/CanvasForm.cs
@@ -96,11 +96,6 @@ namespace TISFAT
 		}
 		#endregion
 
-		public void GLContext_Paint(object sender, PaintEventArgs e)
-		{
-			DrawFrame(Program.MainTimeline.GetCurrentFrame(), false, sender != null);
-		}
-
 		public void DrawFrame(float time, bool render, bool lights)
 		{
 			if (lights)
@@ -153,6 +148,10 @@ namespace TISFAT
 		public void UpdateStickFigurePair()
 		{
 			StickFigurePair = null;
+		}
+		public void GLContext_Paint(object sender, PaintEventArgs e)
+		{
+			DrawFrame(Program.MainTimeline.GetCurrentFrame(), false, sender != null);
 		}
 
 		public void GLContext_MouseDown(object sender, MouseEventArgs e)
